+++ src/components/editor-quill/tests/utils.js
@@ -3,7 +3,7 @@
  * - https://github.com/zenoamaro/react-quill/blob/92beccc417cf0bec16c902c5b307b6a6971344ea/test/utils.js
  */
 import React from 'react'
-import { mount, shallow } from 'enzyme' // eslint-disable-line
+import { mount } from 'enzyme' // eslint-disable-line
 import ReactQuill from 'react-quill'
 import { EditorQuill } from '../'
 
@@ -22,7 +22,7 @@ const sampleData = {
 
 function ReactQuillNode(props, children) {
   props = Object.assign({ // eslint-disable-line
-    modules: {'toolbar': ['underline', 'bold', 'italic']},
+    modules: { 'toolbar': ['underline', 'bold', 'italic'] },
     formats: ['underline', 'bold', 'italic']
   }, props)
 
@@ -69,9 +69,9 @@ function setReactQuillContentsFromHTML(wrapper, html) {
   return editor.clipboard.dangerouslyPasteHTML(html)
 }
 
-function getRenderedEditor(inputHtml = '', onChange = () => {}, blockConfig = undefined) {
+function getRenderedEditor(inputHtml = '', onChange = () => { }, blockConfig = undefined) {
   const blockData = Object.assign(sampleData, { data: { value: inputHtml } })
-  const wrapper = mountEditorQuill({block: blockData, onChange, blockConfig})
+  const wrapper = mountEditorQuill({ block: blockData, onChange, blockConfig })
   return {
     wrapper,
     ReactQuill: wrapper.find(ReactQuill),