+++ src/components/editor-quill/toolbar.js
@@ -1,26 +1,53 @@
 import * as React from 'react'
 import PropTypes from 'prop-types'
 
-// Example Toolbar
-// - http://quilljs.com/standalone/full/
-export const QuillToolbar = function({id}) {
+// based on the example: http://quilljs.com/standalone/full/
+export const QuillToolbar = function({ id }) {
   return (
     <div className="ql-toolbar" id={`toolbar-${id}`} >
-      <select className="ql-header" defaultValue="">
-        <option value="1" />
-        <option value="2" />
-        <option value="3" />
-        <option value="" />
-      </select>
+      <span className="ql-formats">
+        <select className="ql-font" />
+        <select className="ql-size" />
+      </span>
+      <span className="ql-formats">
+        <button className="ql-bold" />
+        <button className="ql-italic" />
+        <button className="ql-underline" />
+        <button className="ql-strike" />
+      </span>
+      <span className="ql-formats">
+        <select className="ql-color" />
+        <select className="ql-background" />
+      </span>
+      <span className="ql-formats">
+        <button className="ql-script" value="sub" />
+        <button className="ql-script" value="super" />
+      </span>
+      <span className="ql-formats">
+        <button className="ql-header" value="1" />
+        <button className="ql-header" value="2" />
+        <button className="ql-blockquote" />
+        <button className="ql-code-block" />
+      </span>
       <span className="ql-formats">
         <button className="ql-list" value="ordered" />
         <button className="ql-list" value="bullet" />
         <button className="ql-indent" value="-1" />
         <button className="ql-indent" value="+1" />
       </span>
-      <button className="ql-bold" />
-      <button className="ql-italic" />
-      <button className="ql-underline" />
+      <span className="ql-formats">
+        <button className="ql-direction" value="rtl" />
+        <select className="ql-align" />
+      </span>
+      <span className="ql-formats">
+        <button className="ql-link" />
+        <button className="ql-image" />
+        <button className="ql-video" />
+        <button className="ql-formula" />
+      </span>
+      <span className="ql-formats">
+        <button className="ql-clean" />
+      </span>
     </div>
   )
 }